Add AppContext and move PubSub to AppContext
