+++ frontend/src/app/ui/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { AuthenticatedApp } from './AuthenticatedApp';
-import { useAuthenticationContext } from '../authentication';
+import { useAuthenticationContext } from '../../authentication';
 import { UnauthenticatedApp } from './UnauthenticatedApp';
 
 export const App: React.FC = () => {
