+++ frontend/src/app/ui/index.ts
@@ -1,0 +1,3 @@
+export * from './App';
+export * from './AuthenticatedApp';
+export * from './UnauthenticatedApp';
