+++ frontend/src/authentication/index.ts
@@ -2,4 +2,5 @@ export * from './entities';
 export * from './clients';
 export * from './interactors';
 export * from './repositories';
+export * from './ui';
 export * from './AuthenticationContext';
