+++ frontend/src/index.tsx
@@ -5,15 +5,19 @@ import './index.css';
 import './i18n';
 import * as serviceWorker from './serviceWorker';
 import { BrowserRouter } from 'react-router-dom';
-import { App } from './app';
+import { App, AppContext, AppProvider } from './app';
 import { AuthenticationProvider } from './authentication';
+
+const appContext = new AppContext();
 
 ReactDOM.render(
   <React.StrictMode>
     <BrowserRouter>
-      <AuthenticationProvider>
-        <App />
-      </AuthenticationProvider>
+      <AppProvider appContext={appContext}>
+        <AuthenticationProvider>
+          <App />
+        </AuthenticationProvider>
+      </AppProvider>
     </BrowserRouter>
   </React.StrictMode>,
   document.getElementById('root')
